Changing relative frequency to decimal instead of int

+++ microservices/Models/WEB_BIOT_SPECIES_GRAB.cs
@@ -20,7 +20,7 @@ namespace microservices.Models
         public string ABUND { get; set; }
         public Nullable<short> contrib_similarity_STRENGTH { get; set; }
         public Nullable<int> SED_ABUND_SED_ABUND { get; set; }
-        public Nullable<short> RELATIVE_FREQUENCY { get; set; }
+        public Nullable<decimal> RELATIVE_FREQUENCY { get; set; }
         public string SORT { get; set; }
     
         public virtual WEB_BIOTOPE WEB_BIOTOPE { get; set; }
